Data.EvalRecord.TH

Documentation

mkEvalRecord :: DecsQ -> DecsQ

Defines a synonym for the bare type with the same name. The strippable definition is suffixed by B, and the covered type is suffixed by H.

mkEvalRecordWith :: DeclareRecordConfig -> DecsQ -> DecsQ

Generate a higher-kinded data declaration using a custom config.

Note that sum types are not currently supported.

data DeclareRecordConfig

Constructors

DeclareRecordConfig

Fields

friends :: [Name]

Members with these types won't be wrapped with 'f @@'

unwrappedName :: String -> Maybe String

generate a type synonym for the Pure version of the type?

baseName :: String -> String

modify the name of the datatype

wrapperName :: Q Name

the name of the type parameter of the wrapper for each field

passthrough :: DeclareRecordConfig

Defines a synonym for the bare type with the same name. The basic definition is suffixed by F, and the unwrapped basic type has no suffix.