Temporal.Bundle

Description

Define and reference workflows and activities using records.

The usage of this module is best explained by example. First, define a record type that contains all of the workflows and activities that you want to define:

data EmailWorkflowsB t f = Workflows
 { emailNewUser :: Wear t f (Int -> Workflow ())
 , sendEmail :: Wear t f (String -> Activity () ())
 , emailPasswordReset :: Wear t f (String -> Workflow ())
 } deriving (Generic)

type EmailWorkflows = EmailWorkflowsB Bare Identity
type EmailWorkflowsH = EmailWorkflowsB Covered

We define all of our fields using the Wear type, which is a type family that lets us convert fields from their direct function definitions into versions that are wrapped in a functor.

Next, we will need some boilerplate instances for this type:

instance FunctorB (EmailWorkflowsB Covered)
instance TraversableB (EmailWorkflowsB Covered)
instance ApplicativeB (EmailWorkflowsB Covered)
instance ConstraintsB (EmailWorkflowsB Covered)
instance BareB EmailWorkflowsB
instance Label (EmailWorkflowsB Covered)

As an alternative to using Generic and deriving these instances, you can use TH to generate them for you. With this style, you don't have to introduce the type parameters manually:

import Temporal.Bundle.TH

passthroughBareB [d|
  data EmailWorkflows = Workflows
    { emailNewUser :: Int -> Workflow ()
    , sendEmail :: String -> Activity () ()
    , emailPasswordReset :: String -> Workflow ()
    }
  |]

This will generate the same instances as the manual version above, with the following type synonyms:

data EmailWorkflowsB = Workflows {..}
type EmailWorkflows = EmailWorkflowsB Bare Identity
type EmailWorkflowsH = EmailWorkflowsB Covered

Now that we have our record type, we can turn it into references and definitions:

Using refs, we can use our record type to provide references to our workflows and activities. References are essentially identifiers that we can use across application / process boundaries to invoke our workflows and activities with Temporal.

wfRefs :: Refs EmailWorkflowsB
wfRefs = refs JSON workflowsImpl

Now, we can implement our workflows and activities:

workflowsImpl :: EmailWorkflows
workflowsImpl = Workflows
  { emailNewUser = userId -> provideCallStack $ do
      h <- startActivity
        wfRefs.sendEmail -- <=== Note how we can reference other workflows and activities using the refs we defined above
        (defaultStartActivityOptions $ ScheduleToClose $ TimeSpec 0 0)
        "foo"
      wait h
  , sendEmail = email -> do
      pure ()
  , emailPasswordReset = email -> do
      pure ()
  }

Lastly, we want a way to register our workflows and activities with the Temporal worker:

workerConfigFromRecord = defsToConfig $ defs JSON workflowsImpl

This will produce a ConfigM that we can use to register our workflows and activities with the Temporal worker as part of a larger ConfigM that we pass to startWorker.

Turning implementations into references and definitions

refs :: CanUseAsRefs r codec => codec -> Refs r

Produce a record of references to workflows and activities from a record of implementations.

The fields in the records produced by this function can be used to invoke the workflows and activities via the Temporal client or within Temporal workflows.

type Refs (f :: (Type -> Type -> Type) -> k) = f Ref

References to workflows and activities that can be used at invokation callsites.

data Ref a b

A reference to a workflow or activity.

Depending on whether the result monad is Workflow or Activity, the Eval instance will resolve to either a KnownWorkflow or KnownActivity.

defs :: CanUseAsDefs f codec env => codec -> Impl f -> Defs env f

Produce a record of definitions from a record of implementations, using the supplied codec.

type Defs env (f :: (Type -> Type -> Type) -> k) = f (Def env)

A wrapped version of the implementations that carries information needed to register workflows and activities with the Temporal worker.

data Def a b c

collectTemporalDefinitions :: forall rec actEnv (f :: Type -> Exp Type). (TraversableRec rec, ConstraintsRec rec, AllRecF (ToDefinitions actEnv) f rec) => rec f -> Definitions actEnv

Convert a record of Def values into a worker and that you can pass to startWorker.

type Impl (f :: (a -> a -> Type) -> k) = f (Pure :: a -> a -> Type)

A bare record type that supplies the actual implementation of workflows and activities.

inWorkflowProxies :: forall synchronicity rec. (RequireCallStack, ConstraintsRec rec, AllRec (ClassF (UseAsInWorkflowProxy synchronicity) Ref) rec, ApplicativeRec rec) => synchronicity -> rec (RefStartOptions <=< Ref) -> rec Ref -> rec (InWorkflowProxies synchronicity <=< Ref)

Combine a record of references with a record of default options to give a record of functions that can be used to start activities and workflows. This is just a convenience function that makes it so that you don't have to call executeActivity or executeChildWorkflow directly as often.

data InWorkflowProxies a b c

type InWorkflowProxyOptions = RefStartOptions <=< Ref

data RefStartOptions a b

class UseAsInWorkflowProxy synchronicity ref where

Methods

useAsInWorkflowProxy :: synchronicity -> ref -> (RefStartOptions @@ ref) -> InWorkflowProxies synchronicity @@ ref

provideDefaultOptions :: (ApplicativeRec rec, ConstraintsRec rec, AllRec FieldToStartOptionDefaults rec) => StartActivityOptions -> StartChildWorkflowOptions -> rec RefStartOptions

Define a record that provides the same default options for all activities in the a record.

This is useful when you want to provide the same default options for all activities in a record. For example, you might want to provide a default timeout or non-retryable error type list for all activities.

class FieldToStartOptionDefaults f where

Methods

refStartOptionsDefaults :: Proxy f -> StartActivityOptions -> StartChildWorkflowOptions -> RefStartOptionsType f

type family RefStartOptionsType f where ...

Equations

RefStartOptionsType (Activity _1 _2) = StartActivityOptions

RefStartOptionsType (_1 -> b) = RefStartOptionsType b

RefStartOptionsType (Workflow _1) = StartChildWorkflowOptions

RefStartOptionsType (KnownWorkflow _1 _2) = StartChildWorkflowOptions

RefStartOptionsType (KnownActivity _1 _2) = StartActivityOptions

Worker management

withTaskQueues :: (TraversableRec rec, MonadUnliftIO m, MonadCatch m) => Client -> WorkerConfigs env rec -> (Workers env rec -> m a) -> m a

linkTaskQueues :: forall m rec env. (TraversableRec rec, MonadUnliftIO m) => Workers env rec -> m ()

startTaskQueues :: (TraversableRec rec, MonadUnliftIO m, MonadCatch m) => Client -> WorkerConfigs env rec -> m (Workers env rec)

Start a Temporal worker for each task queue specified in the MercuryTaskQueues record.

This function starts each worker concurrently, waits for them to initialize, and then returns a worker for each task queue.

shutdownTaskQueues :: forall m rec env. (TraversableRec rec, MonadUnliftIO m) => Workers env rec -> m (rec (ConstFn (Either SomeException ()) :: Type -> Type -> Type))

Stop each Temporal worker for each task queue specified in the MercuryTaskQueues record.

This function stops each worker concurrently, waits for them to complete shutdown (gracefully or not), and then returns.

type WorkerConfigs env (rec :: (b -> Type -> Type) -> k) = rec (ConstFn (WorkerConfig env) :: b -> Type -> Type)

type Workers env (rec :: (b -> Type -> Type) -> k) = rec (ConstFn (Worker env) :: b -> Type -> Type)

Other utilities

coerceRec :: forall (c :: Type -> Exp Type) (d :: Type -> Exp Type) rec. AllRec (Equate c d) rec => rec c -> rec d

class (c @@ a) ~ (d @@ a) => Equate (c :: k1 -> Exp k) (d :: k1 -> Exp k) (a :: k1)

data ProxySync

Constructors

ProxySync

data ProxyAsync

Constructors

ProxyAsync

Reexports

Constraints

type CanUseAsRefs (f :: (Type -> Exp Type) -> Type) codec = (ConstraintsRec f, WitnessFieldTypes f, AllRec (RefFromFunction codec) f, ApplicativeRec f)

type CanUseAsDefs (f :: (Type -> Exp Type) -> Type) codec env = (ConstraintsRec f, WitnessFieldTypes f, AllRec (DefFromFunction codec env) f, ApplicativeRec f)

Constraints needed to turn a record of implementations into a record of definitions.

That is, all of the fields (Workflows, Activities) in the record support the supplied codec, and the activities have an environment of type env.

class RefFromFunction' codec f f => RefFromFunction codec f

class RefFromFunction' codec f original where

Methods

refFromFunction :: Proxy f -> codec -> String -> Proxy original -> Ref @@ original

class DefFromFunction' codec env f f => DefFromFunction codec env f

class DefFromFunction' codec env f original where

Methods

defFromFunction :: Proxy f -> codec -> String -> original -> Def env @@ original

class WorkflowRef f where

Associated Types

type WorkflowArgs f :: [Type]

type WorkflowResult f

Methods

workflowRef :: f -> KnownWorkflow (WorkflowArgs f) (WorkflowResult f)

class ActivityRef f where

Associated Types

type ActivityArgs f :: [Type]

type ActivityResult f

Methods

activityRef :: f -> KnownActivity (ActivityArgs f) (ActivityResult f)

type family InnerActivityResult f where ...

Equations

InnerActivityResult (Activity _1 result) = result

InnerActivityResult (_1 -> b) = InnerActivityResult b

InnerActivityResult a = TypeError ('Text "Expected an Activity, but got " ':<>: 'ShowType a) :: Type

type family ApplyDef f where ...

Equations

ApplyDef (Workflow _1) = WorkflowDefinition

ApplyDef (Activity env _1) = ActivityDefinition env

ApplyDef (_1 -> b) = ApplyDef b

ApplyDef a = TypeError ('Text "Expected a Workflow or Activity, but got " ':<>: 'ShowType a) :: Type

type family ApplyRef original f where ...

Equations

ApplyRef original (Workflow result) = KnownWorkflow (ArgsOf original) result

ApplyRef original (Activity _1 result) = KnownActivity (ArgsOf original) result

ApplyRef original (_1 -> b) = ApplyRef original b

ApplyRef _1 a = TypeError ('Text "Expected a Workflow or Activity, but got " ':<>: 'ShowType a) :: Type