Temporal.Bundle.TH

Description

This module provides Template Haskell functions for generating the boilerplate required to use the Temporal.Bundle module.

This is generally the easiest way to define groups of related activities and workflows with minimal boilerplate.

See Bundle for more example code.

Documentation

temporalBundle :: DecsQ -> DecsQ

Defines a synonym for the bare type with the same name. The strippable definition is suffixed by B, and the covered type is suffixed by H.

data DeclareRecordConfig

Constructors

DeclareRecordConfig

Fields

friends :: [Name]

Members with these types won't be wrapped with 'f @@'

unwrappedName :: String -> Maybe String

generate a type synonym for the Pure version of the type?

baseName :: String -> String

modify the name of the datatype

wrapperName :: Q Name

the name of the type parameter of the wrapper for each field