Temporal.Contrib.OpenTelemetry

Description

OpenTelemetry instrumentation for Temporal workflows and activities.

This module provides an interceptor for Temporal workflows and activities that automatically instruments them with OpenTelemetry tracing. It also provides functions for extracting and injecting OpenTelemetry context from and to Temporal headers.

To use, initialize the OpenTelemetry globalTracerProvider. Then, add the interceptor to your Temporal client and worker configuration.

Documentation

defaultHeaderKey :: Text

"_tracer-data"

data OpenTelemetryInterceptorOptions

Constructors

OpenTelemetryInterceptorOptions

Fields

tracerProvider :: Maybe TracerProvider

headerKey :: Text

defaultOpenTelemetryInterceptorOptions :: OpenTelemetryInterceptorOptions

headersPropagator :: Propagator Context (Map Text Payload) (Map Text Payload)

makeOpenTelemetryInterceptor :: MonadIO m => m (Interceptors env)