Temporal.Duration

Description

Convenient units of time for use with timers and timeouts.

Anything longer than a week (Months, Years) becomes ambiguous because they are variable length. For example, a month can be 28, 29, 30, or 31 days.

If you need to represent a duration as years or months, you may want to use the time library to normalize against a specific time zone and then convert the difference between now and then into a Duration.

Alternatively, consider using Temporal's scheduling features to schedule a workflow to run at one or more specific times in the future.

Documentation

data Duration

A duration of time. Durations are always positive.

durationSeconds :: Duration -> Int64

durationNanoseconds :: Duration -> Int32

mkDuration :: Int64 -> Int32 -> Duration

addDurations :: Duration -> Duration -> Duration

Add two durations together. Durations are subject to integer overflow.

diffTimeToDuration :: DiffTime -> Duration

Convert a DiffTime to a Duration.

N.B. negative DiffTime values will be clamped to 0 (i.e. Duration will be 0) in order to avoid the conversion of negative DiffTime values to extremely large Word64 values.

nominalDiffTimeToDuration :: NominalDiffTime -> Duration

Convert a NominalDiffTime to a Duration.

N.B. negative NominalDiffTime values will be clamped to 0 (i.e. Duration will be 0) in order to avoid the conversion of negative NominalDiffTime values to extremely large Word64 values.

nanoseconds :: Integer -> Duration

Create a Duration from a given number of nanoseconds.

microseconds :: Integer -> Duration

Create a Duration from a given number of microseconds.

milliseconds :: Int64 -> Duration

Create a Duration from a given number of milliseconds.

seconds :: Int64 -> Duration

Create a Duration from a given number of seconds.

minutes :: Int32 -> Duration

Create a Duration from a given number of minutes.

hours :: Int32 -> Duration

Create a Duration from a given number of hours.

days :: Int32 -> Duration

Create a Duration from a given number of days.

weeks :: Int32 -> Duration

Create a Duration from a given number of weeks.

infinity :: Duration

A Duration representing the maximum possible length of time. (Approximately 132 years.)

durationFromProto :: Duration -> Duration

Convert a protocol buffer duration to a Duration.

durationToProto :: Duration -> Duration

Convert a Duration to a protocol buffer duration.

durationToMilliseconds :: Duration -> Double

Convert a Duration to a milliseconds represented as a Double. This is mostly useful for human-readable values emitted by logs, metrics, traces, etc.