Temporal.EphemeralServer

Description

Run an Temporal server in a programatically for testing and development.

Documentation

launchTestServer :: Runtime -> [String] -> IO (Either EphemeralServerError (PortNumber, EphemeralServer))

launchDevServer :: Runtime -> TemporalDevServerConfig -> IO (Either EphemeralServerError EphemeralServer)

withDevServer :: MonadUnliftIO m => Runtime -> TemporalDevServerConfig -> (EphemeralServer -> m a) -> m a

withTestServer :: MonadUnliftIO m => Runtime -> TestServerConfig -> (EphemeralServer -> m a) -> m a

getFreePort :: IO PortNumber

Open a TCP socket, get its port and close the socket. Useful when you have an external service that needs a fresh port.

There is a small race condition present: It's possible to get a free port only for it to be bound by some other process or thread before used

openFreePort :: IO (PortNumber, Socket)

Open a TCP socket on a random free port. This is like warp's openFreePort.

data EphemeralServerError

Constructors

EphemeralServerError ByteString

Core configuration

data EphemeralExeVersion

Constructors

Default SDKDefault
Use a default version for the given SDK name and version.

Fixed String
Specific version.

data EphemeralExe

Constructors

ExistingPath FilePath
Existing path on the filesystem for the executable.

CachedDownload EphemeralExeVersion (Maybe FilePath)
Download the executable if not already there.

data SDKDefault

Constructors

SDKDefault

Fields

sdkName :: String

sdkVersion :: String

data TemporalDevServerConfig

Constructors

TemporalDevServerConfig

Fields

exe :: EphemeralExe

namespace :: FilePath

ip :: String

port :: Maybe Word16

dbFilename :: Maybe FilePath

ui :: Bool

uiPort :: Maybe Word16

log :: (String, String)

extraArgs :: [String]

defaultTemporalDevServerConfig :: TemporalDevServerConfig

data EphemeralServer

shutdownEphemeralServer :: EphemeralServer -> IO (Either ByteString ())

data PortNumber

Port number. Use the Num instance (i.e. use a literal) to create a PortNumber value.

>>> 1 :: PortNumber
1
>>> read "1" :: PortNumber
1
>>> show (12345 :: PortNumber)
"12345"
>>> 50000 < (51000 :: PortNumber)
True
>>> 50000 < (52000 :: PortNumber)
True
>>> 50000 + (10000 :: PortNumber)
60000