Temporal.Exception

Documentation

mkAnnotatedHandlers :: [ApplicationFailureHandler] -> [ApplicationFailureHandler]

prettySrcLoc :: SrcLoc -> String

prettyCallStack :: CallStack -> String

Pretty print a CallStack.

Since: 4.9.0.0

prettyCallStackLines :: CallStack -> [String]

data ActivityCancelReason

A type of exception thrown to a running activity to cancel it due to things happening with the worker, such as a shutdown. This differs from a normal activity cancellation, which uses the cancel function from the async package.

Constructors

NotFound
The activity no longer exists on the server (may already be completed or its workflow may be completed).

CancelRequested
The activity was explicitly cancelled.

Timeout
Activity timeout caused the activity to be marked cancelled.

WorkerShutdown
The worker the activity is running on is shutting down.

UnknownCancellationReason
We received a cancellation reason that we don't know how to handle.

data ActivityFailure

Constructors

ActivityFailure

Fields

message :: Text

activityType :: ActivityType

activityId :: ActivityId

retryState :: RetryState

identity :: Text

cause :: ApplicationFailure

scheduledEventId :: Int64

startedEventId :: Int64

original :: ActivityFailureInfo

stack :: Text

data ApplicationFailure

Constructors

ApplicationFailure

Fields

type' :: Text

message :: Text

nonRetryable :: Bool

details :: [Payload]

stack :: Text

nextRetryDelay :: Maybe Duration

data WorkflowExecutionFailedAttributes

Constructors

WorkflowExecutionFailedAttributes

Fields

failure :: Maybe Failure

retryState :: RetryState

workflowTaskCompletedEventId :: Int64

newExecutionRunId :: Maybe RunId

data CompleteAsync

Asynchronous Activity Completion is a feature that enables an Activity Function to return without causing the Activity Execution to complete. The Temporal Client can then be used to both Heartbeat Activity Execution progress and eventually provide a result.

The intended use-case for this feature is when an external system has the final result of a computation, started by an Activity.

Consider using Asynchronous Activities instead of Signals if the external process is unreliable and might fail to send critical status updates through a Signal.

Consider using Signals as an alternative to Asynchronous Activities to return data back to a Workflow Execution if there is a human in the process loop. The reason is that a human in the loop means multiple steps in the process. The first is the Activity Function that stores state in an external system and at least one other step where a human would “complete” the activity. If the first step fails, you want to detect that quickly and retry instead of waiting for the entire process, which could be significantly longer when humans are involved.

Constructors

CompleteAsync

data ApplicationFailureHandler where

Constructors

ApplicationFailureHandler :: forall e. Exception e => (e -> ApplicationFailure) -> ApplicationFailureHandler

data SomeWorkerException

Constructors

Exception e => SomeWorkerException e

workerExceptionToException :: Exception e => e -> SomeException

workerExceptionFromException :: Exception e => SomeException -> Maybe e

data RuntimeError

Errors that are an issue with the worker itself, not the workflow

These errors should cause the worker to exit, and imply an issue with the SDK itself.

Constructors

RuntimeError String

data WorkflowNotFound

Errors that are the fault of the developer, not the SDK.

Constructors

WorkflowNotFound String

data ActivityNotFound

Constructors

ActivityNotFound String

data QueryNotFound

Constructors

QueryNotFound String

data LogicBugType

Constructors

ReadingCompletionsFailedRun

WorkflowActivationDeadlock

data LogicBug

Constructors

LogicBug LogicBugType

data WorkflowAlreadyStarted

Constructors

WorkflowAlreadyStarted

Fields

workflowAlreadyStartedWorkflowId :: WorkflowId

workflowAlreadyStartedWorkflowType :: WorkflowType

data ChildWorkflowFailed

Constructors

ChildWorkflowFailed Failure

data ChildWorkflowCancelled

Constructors

ChildWorkflowCancelled

data SignalExternalWorkflowFailed

Constructors

SignalExternalWorkflowFailed Failure

data ContinueAsNewException

Constructors

ContinueAsNewException

Fields

continueAsNewWorkflowType :: WorkflowType

continueAsNewArguments :: Vector Payload

continueAsNewOptions :: ContinueAsNewOptions

data AlternativeInstanceFailure

Constructors

AlternativeInstanceFailure

data CancelExternalWorkflowFailed

Constructors

CancelExternalWorkflowFailed Failure

data WorkflowCancelRequested

Constructors

WorkflowCancelRequested

data ActivityCancelled

Constructors

ActivityCancelled Failure

applicationFailureToFailureProto :: ApplicationFailure -> Failure

class ToApplicationFailure e where

Methods

toApplicationFailure :: e -> ApplicationFailure

data SomeApplicationFailure

Constructors

(Exception e, ToApplicationFailure e) => SomeApplicationFailure e

applicationFailureToException :: (Exception e, ToApplicationFailure e) => e -> SomeException

applicationFailureFromException :: Exception e => SomeException -> Maybe e

mkApplicationFailure :: SomeException -> [ApplicationFailureHandler] -> ApplicationFailure

basicHandler :: Exception e => e -> ApplicationFailure

annToPayload :: Annotation -> Payload

annotationHandler :: Exception e => (e -> ApplicationFailure) -> AnnotatedException e -> ApplicationFailure

data NonRetryableError

Constructors

NonRetryableError Bool

data NextRetryDelay

Constructors

NextRetryDelay (Maybe Duration)

standardApplicationFailureHandlers :: [ApplicationFailureHandler]

annotateNonRetryableError :: Annotation

annotateRetryableError :: Annotation

annotateNextRetryDelay :: Duration -> Annotation

annotateNoRetryDelay :: Annotation

data SomeActivityException

Constructors

Exception e => SomeActivityException e

activityExceptionToException :: Exception e => e -> SomeException

activityExceptionFromException :: Exception e => SomeException -> Maybe e

data WorkflowExecutionClosed

Constructors

WorkflowExecutionFailed WorkflowExecutionFailedEventAttributes

WorkflowExecutionTimedOut

WorkflowExecutionCanceled

WorkflowExecutionTerminated

WorkflowExecutionContinuedAsNew

newtype ActivityType

Constructors

ActivityType

Fields

rawActivityType :: Text

data RetryState

Constructors

RetryStateUnspecified

RetryStateInProgress

RetryStateNonRetryableFailure

RetryStateTimeout

RetryStateMaximumAttemptsReached

RetryStateRetryPolicyNotSet

RetryStateInternalServerError

RetryStateCancelRequested