Temporal.Interceptor

Description

Temporal interceptors allow you to customize and extend the behavior of workflows and activities by intercepting and modifying their execution at various points. Interceptors can be used to add cross-cutting concerns, such as logging, security, or monitoring, to your Temporal applications.

This module provides types and functions for defining and composing interceptors for different parts of Temporal workflows and activities.

Types for different types of interceptors:

WorkflowInboundInterceptor: An interceptor for workflows, invoked for commands sent to workflow execution.

WorkflowOutboundInterceptor: An interceptor for workflows, invoked for commands sent from a workflow.

ActivityInboundInterceptor: An interceptor for activities, invoked for commands sent to an activity.

ActivityOutboundInterceptor: An interceptor for activities, invoked for actions executed by an activity.

ClientInterceptors: Interceptors for Temporal clients.

Functions for composing interceptors:

interceptorConvertChildWorkflowHandle: A function for converting child workflow handles within interceptors.

Example usage of interceptors:

Define a custom interceptor for activities using ActivityInboundInterceptor.

myActivityInterceptor :: ActivityInboundInterceptor
myActivityInterceptor = ActivityInboundInterceptor
  { executeActivity = \input next -> do
      -- Perform custom logic before calling the activity function
      result <- next input
      -- Perform custom logic after the activity function completes
      return result
  }

Compose multiple interceptors using Semigroup instances.

combinedInterceptor :: Interceptor
combinedInterceptor = bugsnagInterceptor <> encryptionInterceptor <> otelInterceptor

Provide the interceptors to the Temporal Workflow client and worker.

Documentation

data Interceptors env

Constructors

Interceptors

Fields

workflowInboundInterceptors :: WorkflowInboundInterceptor

workflowOutboundInterceptors :: WorkflowOutboundInterceptor

activityInboundInterceptors :: ActivityInboundInterceptor env

activityOutboundInterceptors :: ActivityOutboundInterceptor env

clientInterceptors :: ClientInterceptors

scheduleClientInterceptors :: ScheduleClientInterceptors

data ExecuteWorkflowInput

Constructors

ExecuteWorkflowInput

Fields

executeWorkflowInputType :: Text

executeWorkflowInputArgs :: Vector Payload

executeWorkflowInputHeaders :: Map Text Payload

executeWorkflowInputInfo :: Info

data WorkflowInboundInterceptor

Constructors

WorkflowInboundInterceptor

Fields

executeWorkflow :: ExecuteWorkflowInput -> (ExecuteWorkflowInput -> IO (WorkflowExitVariant Payload)) -> IO (WorkflowExitVariant Payload)

handleQuery :: HandleQueryInput -> (HandleQueryInput -> IO (Either SomeException Payload)) -> IO (Either SomeException Payload)

data HandleQueryInput

Constructors

HandleQueryInput

Fields

handleQueryId :: Text

handleQueryInputType :: Text

handleQueryInputArgs :: Vector Payload

handleQueryInputHeaders :: Map Text Payload

data WorkflowOutboundInterceptor

Constructors

WorkflowOutboundInterceptor

Fields

scheduleActivity :: ActivityInput -> (ActivityInput -> IO (Task Payload)) -> IO (Task Payload)

startChildWorkflowExecution :: Text -> StartChildWorkflowOptions -> (Text -> StartChildWorkflowOptions -> IO (ChildWorkflowHandle Payload)) -> IO (ChildWorkflowHandle Payload)

continueAsNew :: forall a. Text -> ContinueAsNewOptions -> (Text -> ContinueAsNewOptions -> IO a) -> IO a

data ActivityInput

Constructors

ActivityInput

Fields

activityType :: Text

args :: Vector Payload

options :: StartActivityOptions

seq :: Sequence

data ActivityInboundInterceptor env

Constructors

ActivityInboundInterceptor

Fields

executeActivity :: env -> ExecuteActivityInput -> (env -> ExecuteActivityInput -> IO (Either String Payload)) -> IO (Either String Payload)

data ExecuteActivityInput

Constructors

ExecuteActivityInput

Fields

activityArgs :: Vector Payload

activityHeaders :: Map Text Payload

activityInfo :: ActivityInfo

data ActivityOutboundInterceptor env

Constructors

ActivityOutboundInterceptor

data WorkflowExitVariant a

Constructors

WorkflowExitContinuedAsNew ContinueAsNewException

WorkflowExitCancelled WorkflowCancelRequested

WorkflowExitFailed SomeException

WorkflowExitSuccess a

data ClientInterceptors

Constructors

ClientInterceptors

Fields

start :: WorkflowType -> WorkflowId -> StartWorkflowOptions -> Vector Payload -> (WorkflowType -> WorkflowId -> StartWorkflowOptions -> Vector Payload -> IO (WorkflowHandle Payload)) -> IO (WorkflowHandle Payload)

queryWorkflow :: QueryWorkflowInput -> (QueryWorkflowInput -> IO (Either QueryRejected Payload)) -> IO (Either QueryRejected Payload)

signalWithStart :: SignalWithStartWorkflowInput -> (SignalWithStartWorkflowInput -> IO (WorkflowHandle Payload)) -> IO (WorkflowHandle Payload)

data QueryWorkflowInput

Constructors

QueryWorkflowInput

Fields

queryWorkflowWorkflowId :: WorkflowId

queryWorkflowRejectCondition :: QueryRejectCondition

queryWorkflowHeaders :: Map Text Payload

queryWorkflowType :: Text

queryWorkflowRunId :: Maybe RunId

queryWorkflowArgs :: Vector Payload

data SignalWithStartWorkflowInput

Constructors

SignalWithStartWorkflowInput

Fields

signalWithStartWorkflowType :: WorkflowType

signalWithStartWorkflowId :: WorkflowId

signalWithStartArgs :: Vector Payload

signalWithStartSignalName :: Text

signalWithStartSignalArgs :: Vector Payload

signalWithStartOptions :: StartWorkflowOptions

data ScheduleClientInterceptors

Constructors

ScheduleClientInterceptors

Fields

scheduleWorkflowAction :: StartWorkflowOptions -> [Payload] -> IO StartWorkflowOptions

Unlike the other interceptors, this one is invoked on the construction of the workflow start options

interceptorConvertChildWorkflowHandle :: ChildWorkflowHandle a -> (a -> IO b) -> ChildWorkflowHandle b

This is only intended for use by interceptors. Normal workflow code should be able to use the fmap instance for simple transformations or else provide an appropriate codec.