Temporal.Operator

Description

This module provides bindings for the Temporal Operator API. The Operator API is used to manage Temporal namespaces, clusters, and other administrative tasks.

Documentation

data IndexedValueType

Constructors

Text

Keyword

Int

Double

Bool

Datetime

KeywordList

UnrecognizedIndexedValueType

data SearchAttributes

Constructors

SearchAttributes

Fields

customAttributes :: Map SearchAttributeKey IndexedValueType

systemAttributes :: Map SearchAttributeKey IndexedValueType

storageSchema :: Map Text Text

searchAttributeTypeFromProto :: IndexedValueType -> IndexedValueType

searchAttributeTypeToProto :: IndexedValueType -> IndexedValueType

listSearchAttributes :: MonadIO m => Client -> Namespace -> m (Either RpcError SearchAttributes)

addSearchAttributes :: MonadIO m => Client -> Namespace -> Map SearchAttributeKey IndexedValueType -> m (Either RpcError ())