Temporal.Payload

Description

A Payload represents binary data such as input and output from Activities and Workflows. Payloads contain metadata that describe the binary data, such as its data type or other arbitrary values for use by custom encoders/converters.

When processed through the SDK, the default Data Converter serializes your data/value to a Payload before sending it to the Temporal Server. The default Data Converter processes supported type values to Payloads, and you can create a custom Payload Converter to convert your custom object types.

You can additionally apply custom codecs (such as for encryption or compression) on your Payloads to wrap them into new encoded Payloads.

Documentation

data Payload

Constructors

Payload

Fields

payloadData :: ByteString

payloadMetadata :: Map Text ByteString

class Codec fmt a where

We want to be able to serialize & deserialize the parameters of a function using an arbitrary serialization format.

Minimal complete definition

encoding, encode, decode

Methods

encoding :: fmt -> Proxy a -> ByteString

Similar to a content-type header, this is a string that identifies the format of the payload. it will be set on the encoding metadata field of the payload.

messageType :: fmt -> a -> ByteString

default messageType :: Typeable a => fmt -> a -> ByteString

encode :: fmt -> a -> IO Payload

decode :: fmt -> Payload -> IO (Either String a)

data JSON

Constructors

JSON

data Null

Constructors

Null

data Binary

Direct binary serialization.

A generalized instance like the one for JSON is not provided because there are many possible binary serialization formats and libraries.

You can provide a simple version of this instance for your own codebase using something like the cereal package:

import Data.Serialize (Serialize)
import qualified Data.Serialize as Serialize
instance {\-# OVERLAPPABLE #-\} (Typeable a, Serialize a) => Codec Binary a where
  encodingType _ _ = "binary/plain"
  encodePayload _ x = Serialize.encode x
  decode _ = Serialize.decode . inputPayloadData

Constructors

Binary

data Protobuf

Constructors

Protobuf

data Zlib

Constructors

Zlib

Fields

minimumEncodingSize :: Int

Skip compression when the size of the response body is below this amount of bytes (default: 860.)

Setting this option to less than 150 will actually increase the size of outgoing data if its original size is less than 150 bytes.

compressParams :: CompressParams

decompressParams :: DecompressParams

zlib :: Zlib

applyPayloads :: ApplyPayloads codec args => codec -> Proxy args -> Proxy result -> (args :->: result) -> Vector Payload -> IO (Either String result)

data Around base

A codec that uses the base codec for actually encoding and decoding, but supports additional pre- and post-processing. This is useful for adding compression, encryption, or other transformations to an existing codec.

Constructors

Around

Fields

around :: PayloadProcessor

baseCodec :: base

mkPayloadProcessor :: Codec fmt Payload => fmt -> PayloadProcessor

addPayloadProcessor :: Codec fmt Payload => fmt -> PayloadProcessor -> PayloadProcessor

data PayloadProcessor

A codec that post-processes the payload after encoding and pre-processes before decoding.

Combining multiple PayloadProcessors will apply them in the order that they are added. For example, if you want to compress and then encrypt, you would add the compression processor first, then the encryption processor.

Constructors

PayloadProcessor

Fields

payloadProcessorEncode :: Payload -> IO Payload

payloadProcessorDecode :: Payload -> IO (Either String Payload)

class ApplyPayloads codec (args :: [Type])

Minimal complete definition

applyPayloads

type family ArgsOf f :: [Type] where ...

Equations

ArgsOf (arg -> rest) = arg ': ArgsOf rest

ArgsOf _1 = '[] :: [Type]

type family ResultOf (m :: Type -> Type) f where ...

Equations

ResultOf m (_1 -> rest) = ResultOf m rest

ResultOf m (m result) = result

ResultOf m result = TypeError ((('Text "This function must use the (" ':<>: 'ShowType m) ':<>: 'Text ") monad.") ':$$: ('Text "Current type: " ':<>: 'ShowType result)) :: Type

type family (args :: [Type]) :->: result where ... infixr 0

Construct a function type from a list of argument types and a result type.

Equations

('[] :: [Type]) :->: result = result

(arg ': args) :->: result = arg -> args :->: result

convertToProtoPayload :: Payload -> Payload

convertFromProtoPayload :: Payload -> Payload

Vararg manipulation

class VarArgs (args :: [Type]) where

Methods

sequenceArgs :: Monad m => Proxy (m result) -> m (args :->: m result) -> args :->: m result

mapResult :: (result -> result') -> (args :->: result) -> args :->: result'

foldlArgs :: AllArgs c args => Proxy c -> (forall a. c a => b -> a -> b) -> b -> args :->: b

hoistResult :: forall {k} (args :: [Type]) (result :: k) m n. VarArgs args => (forall (x :: k). m x -> n x) -> (args :->: m result) -> args :->: n result

foldMapArgs :: forall (args :: [Type]) c m. (VarArgs args, AllArgs c args, Monoid m) => (forall a. c a => a -> m) -> args :->: m

foldMArgs :: forall (args :: [Type]) c b m. (VarArgs args, AllArgs c args, Monad m) => (forall a. c a => b -> a -> m b) -> b -> args :->: m b

foldMArgs_ :: forall (args :: [Type]) c b m. (VarArgs args, AllArgs c args, Monad m) => (forall a. c a => b -> a -> m b) -> b -> args :->: m ()

withArgs :: forall (args :: [Type]) result codec. (VarArgs args, AllArgs (Codec codec) args) => codec -> (Vector UnencodedPayload -> result) -> args :->: result

type UnencodedPayload = IO Payload

processorEncodePayloads :: (MonadIO m, Traversable f) => PayloadProcessor -> f Payload -> m (f Payload)

This is the final step in the process of encoding payloads.

It takes a list of suspended payload encodings and runs them through the payload processor. This allows us to

processorDecodePayloads :: (MonadIO m, Traversable f) => PayloadProcessor -> f Payload -> m (f Payload)

type family AllArgs (c :: Type -> Constraint) (args :: [Type]) where ...

Applies a constraint to all types in a type-level list.

Equations

AllArgs _1 ('[] :: [Type]) = ()

AllArgs c (arg ': args) = (c arg, AllArgs c args)

type GatherArgs codec (args :: [Type]) = (VarArgs args, AllArgs (Codec codec) args)

type GatherArgsOf codec f = GatherArgs codec (ArgsOf f)

type FunctionSupportsCodec codec (args :: [Type]) result = (GatherArgs codec args, Codec codec result, Typeable result, ApplyPayloads codec args)

type FunctionSupportsCodec' (m :: Type -> Type) codec f = (FunctionSupportsCodec codec (ArgsOf f) (ResultOf m f), f ~ (ArgsOf f :->: m (ResultOf m f)))

data ValueError

Used to denote that a payload either failed to encode or decode

Constructors

ValueError String