Temporal.Replay

Description

Temporal.Replay supports replaying a workflow from its event history. Use for troubleshooting and backwards compatibility unit tests.

For example, if a workflow fails, in production, then its history can be downloaded through the UI or CLI and replayed in a debugger as many times as necessary.

Use newReplayWorker to create unit tests that check if workflow changes are backwards compatible. It is important to maintain backwards compatibility to ensure that new deployments are not going to break open workflows.

Documentation

readHistoryProtobufFile :: MonadIO m => FilePath -> m History

writeHistoryProtobufFile :: FilePath -> History -> IO ()