Temporal.TH

Description

The TH module provides Template Haskell-based utilities for defining, configuring, and registering Temporal Workflows and Activities in a type-safe and declarative manner. This module simplifies the process of integrating Temporal with Haskell applications by automating much of the boilerplate code typically required for setting up Workflows and Activities.

Key features of this module include:

Automatic registration of Workflows and Activities.

Type-safe configuration options for both Workflows and Activities.

Generation of necessary typeclass instances for Workflow and Activity functions.

Support for custom naming and aliasing of Activities.

Integration with Haskell's strong type system to ensure correctness at compile-time.

The main functions provided by this module are:

registerWorkflow and registerWorkflowWithOptions: For registering Workflow functions.

registerActivity and registerActivityWithOptions: For registering Activity functions.

discoverDefinitions: For automatically discovering and collecting all defined Workflows and Activities.

When you apply the registration functions to your Workflow and Activity functions, a number of things happen automatically:

The registration functions generate a new data type with the same name as your Workflow or Activity function (capitalized). Example: If you define a Workflow function named myWorkflow, the generated data type will be named MyWorkflow.

A number of type classes are generated or derived for your Workflow or Activity function to support invoking them.

Example usage:

{-# LANGUAGE DerivingVia #-}
{-# LANGUAGE TemplateHaskell #-}
{-# LANGUAGE ImportQualifiedPost #-}

module MyWorkflows where

import Language.Haskell.TH (newDeclarationGroup)
import RequireCallStack (provideCallStack)
import Temporal.Activity
import Temporal.TH
import Temporal.TH.Options
import Temporal.Workflow
import Temporal.Client qualified as Client

-- Simple activity registration
actWithoutTimeoutDefault :: Activity () ()
actWithoutTimeoutDefault = pure ()
$(registerActivity 'actWithoutTimeoutDefault)

-- Activity registration with name override
actWithNameOverride :: Activity () ()
actWithNameOverride = pure ()
$(registerActivityWithOptions 'actWithNameOverride defaultActivityConfiguration
  { actNameOverride = Just "my-name"
  })

-- Activity registration with aliases
actWithAliases :: Activity () ()
actWithAliases = pure ()
$(registerActivityWithOptions 'actWithAliases defaultActivityConfiguration
  { actAliases = ["alias1", "alias2"]
  })

-- Workflow registration
myWorkflow :: Int -> Workflow ()
myWorkflow _ = pure ()
$(registerWorkflow 'myWorkflow)

-- Use newDeclarationGroup to separate Template Haskell splices
$(newDeclarationGroup)

-- Workflow referencing a child workflow
workflowReferencingChild :: Workflow ()
workflowReferencingChild = provideCallStack do
  _ <- startChildWorkflow MyWorkflow defaultChildWorkflowOptions 0
  executeChildWorkflow MyWorkflow defaultChildWorkflowOptions 1

-- Activity referencing a workflow
actionReferencingWorkflow :: Activity () ()
actionReferencingWorkflow = provideCallStack do
  _ <- Client.start MyWorkflow "1" (Client.startWorkflowOptions (TaskQueue "foo")) 1
  Client.execute MyWorkflow "2" (Client.startWorkflowOptions (TaskQueue "foo")) 1

Note the use of newDeclarationGroup from Language.Haskell.TH to separate Template Haskell splices, which can be necessary in some cases to avoid compilation errors– namely, when you want to invoke a Workflow or Activity later in the same module.

The one use case where the TH mechanism struggles is when you want to use the continueAsNew function from Temporal.Workflow in a Workflow to start a new instance of the same Workflow. In this case, you can define the machinery manually:

workflowThatContinuesAsNew :: Int -> Workflow ()
workflowThatContinuesAsNew x = do
  if x < 1000
  then continueAsNew WorkflowThatContinuesAsNew defaultContinueAsNewOptions (x + 1)
  else return ()

data WorkflowThatContinuesAsNew = WorkflowThatContinuesAsNew
  deriving anyclass (WorkflowFn)
  deriving (WorkflowDef, WorkflowRef) via WorkflowImpl WorkflowThatContinuesAsNew

instance Fn WorkflowThatContinuesAsNew where
  fnName = Text.pack $ show 'workflowThatContinuesAsNew
  fnDefinition _ = provideCallStack workflowThatContinuesAsNew
  fnSing = WorkflowThatContinuesAsNew

Documentation

discoverDefinitions :: (RequireCallStack, Typeable env) => [SomeDict WorkflowFn] -> [SomeDict ActivityFn] -> Definitions env

type SomeDict = SomeDictOf (Proxy :: k -> Type)

In the case where you only need evidence that the dictionary exists, then you can use this type. By carrying a Proxy instead of a real value, we can summon these up whenever we have an instance of the type in question.

This is useful when you have a type class that specifies some *static* behavior that is useful, or a type class that can provide means of creatingretrievingworking with data. Consider the PersistEntity class from the persistent database library. An instance of that class can be used to get the EntityDef that describes how the type interacts with the database, or you can even selectList and grab all entities out of the database.

someEntity :: SomeDict PersistEntity
someEntity = someDict @User

With this value, we can now extract the EntityDef:

someEntityDef :: EntityDef
someEntityDef =
    case someEntity of
        SomeDictOf (Proxy :: Proxy a) ->
            entityDef (Proxy :: Proxy a)

We can also load all the rows out of the database.

verifyRowsCanLoad :: SqlPersistT IO ()
verifyRowsCanLoad =
    case someEntity of
        SomeDictOf (Proxy :: Proxy a) -> do
            rows <- selectList [] [] :: SqlPersistT IO [Entity a]
            mapM_ (print . entityKey) rows

Since: some-dict-of-0.1.0.0

data SomeDictOf (f :: k -> Type) (c :: k -> Constraint) where

A datatype that carries evidence of type class membership for some type, along with a datatype indexed by that type. This is confusing, so let's look at some examples.

The alias SomeDict clazz uses Proxy for the datatype index. This means that the wrapper is a Proxy :: Proxy a, and you know that the type a has an instance of clazz a. The SomeDictOf type does not actually *carry* any values of this type.

That is not to say that we *necessarily* won't have a value - consider SomeDictOf Identity, which actually does hold a value. We can use this to make an existential Show wrapper.

showSomeDict :: SomeDictOf Identity Show
showSomeDict =
    SomeDictOf (Identity 3 :: Identity Int)

We can happily have a [SomeDictOf Identity Show], or a Map String (SomeDictOf Identity 'Show), or similar.

Constructing them is easy enough. Consuming them can be a bit trickier. Let's look at a SomeDict Monoid.

monoid :: SomeDictOf Proxy Monoid
monoid =
    SomeDictOf (Proxy :: Proxy Text)

All we know about this is that it's carrying a datatype with a Monoid instance. We'll case-match on the value, and then in the case branch, we'll have evidence that (whatever the underlying type is) that it has a Monoid instance.

We'll repackage the value, but instead of using Proxy, we'll stuff mempty into Identity.

useMonoid :: SomeDictOf Proxy Monoid -> SomeDictOf Identity Monoid
useMonoid someDictOfProxy =
    case someDictOfProxy of
        SomeDictOf (Proxy :: Proxy a) ->
            SomeDictOf (Identity (mempty :: a))

Since: some-dict-of-0.1.0.0

Constructors

SomeDictOf :: forall {k} (c :: k -> Constraint) (a :: k) (f :: k -> Type). c a => f a -> SomeDictOf f c

registerActivity :: (Quote m, Quasi m) => Name -> m [Dec]

registerActivityWithOptions :: forall codec m. (Quote m, Quasi m, Lift codec) => Name -> ActivityConfig codec -> m [Dec]

data ActivityConfig c

Constructors

ActivityConfig

Fields

activityConfigNameOverride :: Maybe Text

activityConfigAliases :: [Text]

activityConfigCodec :: c

defaultActivityConfig :: ActivityConfig JSON

registerWorkflow :: (Quote m, Quasi m) => Name -> m [Dec]

registerWorkflowWithOptions :: forall codec m. (Quote m, Quasi m, Lift codec) => Name -> WorkflowConfig codec -> m [Dec]

data WorkflowConfig codec

Constructors

WorkflowConfig

Fields

workflowConfigNameOverride :: Maybe Text

workflowConfigAliases :: [Text]

workflowConfigCodec :: codec

defaultWorkflowConfig :: WorkflowConfig JSON

class Fn t where

Associated Types

type FnType t

Methods

fnName :: t -> Text

fnDefinition :: t -> RequireCallStack => FnType t

fnSing :: t

class Fn f => WorkflowFn f where

Minimal complete definition

Nothing

Associated Types

type WorkflowCodec f

type WorkflowCodec f = JSON

Methods

workflowImpl :: f -> ProvidedWorkflow (FnType f)

default workflowImpl :: (RequireCallStack, FnType f ~ (ArgsOf (FnType f) :->: Workflow (ResultOf Workflow (FnType f))), FunctionSupportsCodec (WorkflowCodec f) (ArgsOf (FnType f)) (ResultOf Workflow (FnType f))) => f -> ProvidedWorkflow (FnType f)

workflowConfig :: f -> WorkflowConfig (WorkflowCodec f)

default workflowConfig :: WorkflowCodec f ~ JSON => f -> WorkflowConfig (WorkflowCodec f)

class WorkflowRef f where

Associated Types

type WorkflowArgs f :: [Type]

type WorkflowResult f

Methods

workflowRef :: f -> KnownWorkflow (WorkflowArgs f) (WorkflowResult f)

class WorkflowDef a where

Methods

workflowDefinition :: a -> WorkflowDefinition

class (Fn f, Typeable (FnActivityEnv (FnType f))) => ActivityFn f where

Minimal complete definition

Nothing

Associated Types

type ActivityCodec f

type ActivityCodec f = JSON

Methods

activityEnvType :: f -> TypeRep

activityImpl :: f -> ProvidedActivity (FnActivityEnv (FnType f)) (FnType f)

default activityImpl :: (RequireCallStack, env ~ FnActivityEnv (FnType f), FnType f ~ (ArgsOf (FnType f) :->: Activity env (ResultOf (Activity env) (FnType f))), FunctionSupportsCodec (ActivityCodec f) (ArgsOf (FnType f)) (ResultOf (Activity env) (FnType f))) => f -> ProvidedActivity (FnActivityEnv (FnType f)) (FnType f)

activityConfig :: f -> ActivityConfig (ActivityCodec f)

default activityConfig :: ActivityCodec f ~ JSON => f -> ActivityConfig (ActivityCodec f)

class ActivityRef f where

Associated Types

type ActivityArgs f :: [Type]

type ActivityResult f

Methods

activityRef :: f -> KnownActivity (ActivityArgs f) (ActivityResult f)

class ActivityDef a where

Associated Types

type ActivityDefinitionEnv a

Methods

activityDefinition :: a -> ActivityDefinition (ActivityDefinitionEnv a)

fnSingE :: (Quote m, Quasi m) => Name -> m Exp

fnSingDataAndConName :: Name -> Name

newtype ActivityImpl f

Constructors

ActivityImpl f

newtype WorkflowImpl f

Constructors

WorkflowImpl f

bringRegisteredTemporalFunctionsIntoScope :: Q [Dec]

Alias for newDeclarationGroup that makes the intent a bit more clear.