Temporal.Workflow

Description

This module provides the core functionality for defining functions that can be executed as Temporal workflows.

In day-to-day conversations, the term Workflow frequently denotes either a Workflow Type, a Workflow Definition, or a Workflow Execution.

What is a Workflow Definition?

A Workflow Definition is the code that defines the constraints of a Workflow Execution.

A Workflow Definition is often also referred to as a Workflow Function. In Temporal's documentation, a Workflow Definition refers to the source for the instance of a Workflow Execution, while a Workflow Function refers to the source for the instance of a Workflow Function Execution.

A Workflow Execution effectively executes once to completion, while a Workflow Function Execution occurs many times during the life of a Workflow Execution.

Deterministic constraints

A critical aspect of developing Workflow Definitions is ensuring they exhibit certain deterministic traits – that is, making sure that the same Commands are emitted in the same sequence, whenever a corresponding Workflow Function Execution (instance of the Function Definition) is re-executed.

The execution semantics of a Workflow Execution include the re-execution of a Workflow Function, which is called a Replay. The use of Workflow APIs in the function is what generates Commands. Commands tell the Cluster which Events to create and add to the Workflow Execution's Event History. When a Workflow Function executes, the Commands that are emitted are compared with the existing Event History. If a corresponding Event already exists within the Event History that maps to the generation of that Command in the same sequence, and some specific metadata of that Command matches with some specific metadata of the Event, then the Function Execution progresses.

For example, using an SDK's "Execute Activity" API generates the ScheduleActivityTask Command. When this API is called upon re-execution, that Command is compared with the Event that is in the same location within the sequence. The Event in the sequence must be an ActivityTaskScheduled Event, where the Activity name is the same as what is in the Command.

If a generated Command doesn't match what it needs to in the existing Event History, then the Workflow Execution returns a non-deterministic error.

The following are the two reasons why a Command might be generated out of sequence or the wrong Command might be generated altogether:

Code changes are made to a Workflow Definition that is in use by a running Workflow Execution.

There is intrinsic non-deterministic logic (such as inline random branching).

Code changes can cause non-deterministic behavior

The Workflow Definition can change in very limited ways once there is a Workflow Execution depending on it. To alleviate non-deterministic issues that arise from code changes, we recommend using Workflow Versioning.

For example, let's say we have a Workflow Definition that defines the following sequence:

Start and wait on a Timer/sleep.

Spawn and wait on an Activity Execution.

Complete.

We start a Worker and spawn a Workflow Execution that uses that Workflow Definition. The Worker would emit the StartTimer Command and the Workflow Execution would become suspended.

Before the Timer is up, we change the Workflow Definition to the following sequence:

Spawn and wait on an Activity Execution.

Start and wait on a Timer/sleep.

Complete.

When the Timer fires, the next Workflow Task will cause the Workflow Function to re-execute. The first Command the Worker sees would be ScheduleActivityTask Command, which wouldn't match up to the expected TimerStarted Event.

The Workflow Execution would fail and return a non-deterministic error.

The following are examples of minor changes that would not result in non-determinism errors when re-executing a History which already contain the Events:

Changing the duration of a Timer

Changing the arguments to:

The Activity Options in a call to spawn an Activity Execution (local or nonlocal).

The Child Workflow Options in a call to spawn a Child Workflow Execution.

Call to Signal an External Workflow Execution.

Adding a Signal Handler for a Signal Type that has not been sent to this Workflow Execution.

Workflow Versioning

Note: not yet implemented

The Workflow Versioning feature enables the creation of logical branching inside a Workflow Definition based on a developer specified version identifier. This feature is useful for Workflow Definition logic needs to be updated, but there are running Workflow Executions that currently depends on it. It is important to note that a practical way to handle different versions of Workflow Definitions, without using the versioning API, is to run the different versions on separate Task Queues.

Handling unreliable Worker Processes

You do not handle Worker Process failure or restarts in a Workflow Definition.

Workflow Function Executions are completely oblivious to the Worker Process in terms of failures or downtime. The Temporal Platform ensures that the state of a Workflow Execution is recovered and progress resumes if there is an outage of either Worker Processes or the Temporal Cluster itself. The only reason a Workflow Execution might fail is due to the code throwing an error or exception, not because of underlying infrastructure outages.

Documentation

data Workflow a

The Workflow monad is a constrained execution environment that helps developers write code that can be executed deterministically and reliably.

If a Workflow execution is interrupted, for example due to a server failure, or otherwise fails to complete, the Temporal service will automatically replay the Workflow execution up to the point of interruption with the same inputs at each step in the function.

The st state may be used to store information that is needed to respond to any queries or signals that are received by the Workflow execution.

Workflow code must be deterministic. This means:

no threading

no randomness

no external calls to processes

no network I/O

no global state mutation

no system date or time

This might seem like a lot of restrictions, but Temporal provides a number of functions that allow you to use similar functionality in a deterministic way.

A critical aspect of developing Workflow Definitions is ensuring they exhibit certain deterministic traits – that is, making sure that the same Commands are emitted in the same sequence, whenever a corresponding Workflow Function Execution (instance of the Function Definition) is re-executed.

data WorkflowDefinition

This is a Workflow function that can be called from the outside world.

You can create a definition via provideWorkflow.

To make a worker use this definition, you must add it to the WorkerConfig via addWorkflow.

Constructors

WorkflowDefinition

Fields

workflowName :: Text

workflowRun :: Vector Payload -> IO (Either String (Workflow Payload))

data KnownWorkflow (args :: [Type]) result

A KnownWorkflow is a handle that contains all the information needed to start a Workflow either as a child workflow or as a top-level workflow via a Client.

Constructors

FunctionSupportsCodec codec args result => KnownWorkflow

Fields

knownWorkflowCodec :: codec

knownWorkflowName :: Text

data ProvidedWorkflow f

Constructors

ProvidedWorkflow

Fields

definition :: WorkflowDefinition

reference :: KnownWorkflow (ArgsOf f) (ResultOf Workflow f)

provideWorkflow :: forall codec f. (f ~ (ArgsOf f :->: Workflow (ResultOf Workflow f)), FunctionSupportsCodec codec (ArgsOf f) (ResultOf Workflow f)) => codec -> Text -> (RequireCallStackImpl => f) -> ProvidedWorkflow f

A utility function for constructing a WorkflowDefinition from a function as well as a KnownWorkflow value. This is useful for keeping the argument, codec, and result types in sync with each other so that changes to the function are reflected at their use sites.

myWorkflow :: Workflow () () Int
myWorkflow = pure 1

myWorkflowDef :: ProvidedWorkflow () () Int
myWorkflowDef = provideWorkflow
  JSON -- codec
  @"myWorkflow" -- visible name of the workflow
  () -- initial state
  myWorkflow -- the workflow function

data Task a

An async action handle that can be awaited or cancelled.

newtype StartToClose

An Activity Execution must have either this timeout (Start-To-Close) or the Schedule-To-Close Timeout set. We recommend always setting this timeout; however, make sure that Start-To-Close Timeout is always set to be longer than the maximum possible time for the Activity Execution to complete. For long running Activity Executions, we recommend also using Activity Heartbeats and Heartbeat Timeouts.

Tip: We strongly recommend setting a Start-To-Close Timeout.

The Temporal Server doesn't detect failures when a Worker loses communication with the Server or crashes. Therefore, the Temporal Server relies on the Start-To-Close Timeout to force Activity retries.

Constructors

StartToClose Duration

newtype ScheduleToClose

A Schedule-To-Close Timeout is the maximum amount of time allowed for the overall Activity Execution, from when the first Activity Task is scheduled to when the last Activity Task, in the chain of Activity Tasks that make up the Activity Execution, reaches a Closed status.

An Activity Execution must have either this timeout (Schedule-To-Close) or Start-To-Close set. By default, an Activity Execution Retry Policy dictates that retries occur for up to 10 years. This timeout can be used to control the overall duration of an Activity Execution in the face of failures (repeated Activity Task Executions), without altering the Maximum Attempts field of the Retry Policy.

Constructors

ScheduleToClose Duration

data These a b

The These type represents values with two non-exclusive possibilities.

This can be useful to represent combinations of two values, where the combination is defined if either input is. Algebraically, the type These A B represents (A + B + AB), which doesn't factor easily into sums and products--a type like Either A (B, Maybe A) is unclear and awkward to use.

These has straightforward instances of Functor, Monad, &c., and behaves like a hybrid error/writer monad, as would be expected.

For zipping and unzipping of structures with These values, see Data.Align.

Constructors

This a

That b

These a b

Workflow monad operations

Activity operations

An Activity is an IO-based function that executes a single, well-defined action (either short or long running), such as calling another service, transcoding a media file, or sending an email message. Activity code can be non-deterministic.

We recommend that it be idempotent.

Workflow code orchestrates the execution of Activities, persisting the results. If an Activity Function Execution fails, any future execution starts from initial state (except Heartbeats).

Activity Functions are executed by Worker Processes. When the Activity Function returns, the Worker sends the results back to the Temporal Cluster as part of the ActivityTaskCompleted Event. The Event is added to the Workflow Execution's Event History.

class ActivityRef f where

Associated Types

type ActivityArgs f :: [Type]

type ActivityResult f

Methods

activityRef :: f -> KnownActivity (ActivityArgs f) (ActivityResult f)

data KnownActivity (args :: [Type]) result

Constructors

FunctionSupportsCodec codec args result => KnownActivity

Fields

knownActivityCodec :: codec

knownActivityName :: Text

data StartActivityOptions

Constructors

StartActivityOptions

Fields

activityId :: Maybe ActivityId

The identifier for an Activity Execution. The identifier can be generated by the system, or it can be provided by the Workflow code that spawns the Activity Execution. The identifier is unique among the open Activity Executions of a Workflow Run. (A single Workflow Run may reuse an Activity Id if an earlier Activity Execution with the same Id has closed.)

An Activity Id can be used to complete the Activity asynchronously.

taskQueue :: Maybe TaskQueue

timeout :: ActivityTimeoutPolicy

scheduleToStartTimeout :: Maybe Duration

A Schedule-To-Start Timeout is the maximum amount of time that is allowed from when an Activity Task is scheduled (that is, placed in a Task Queue) to when a Worker starts (that is, picks up from the Task Queue) that Activity Task. In other words, it's a limit for how long an Activity Task can be enqueued.

heartbeatTimeout :: Maybe Duration

A Heartbeat Timeout is the maximum time between Activity Heartbeats. If this timeout is reached, the Activity Task fails and a retry occurs if a Retry Policy dictates it.

retryPolicy :: Maybe RetryPolicy

A Retry Policy works in cooperation with the timeouts to provide fine controls to optimize the execution experience.

A Retry Policy is a collection of attributes that instructs the Temporal Server how to retry a failure of a Workflow Execution or an Activity Task Execution. (Retry Policies do not apply to Workflow Task Executions, which always retry indefinitely.)

cancellationType :: ActivityCancellationType

headers :: Map Text Payload

disableEagerExecution :: Bool

If true, will disable eager activity execution. Eager activity execution is an optimization on some servers that sends activities back to the same worker as the calling workflow if they can run there. This setting is experimental and may be removed in a future release.

data ActivityCancellationType

Defines how the workflow will wait (or not) for cancellation of the activity to be confirmed.

Constructors

ActivityCancellationTryCancel
Initiate a cancellation request and immediately report cancellation to the workflow.

ActivityCancellationWaitCancellationCompleted
Wait for activity cancellation completion. Note that activity must heartbeat to receive a cancellation notification. This can block the cancellation for a long time if activity doesn't heartbeat or chooses to ignore the cancellation request.

ActivityCancellationAbandon
Do not request cancellation of the activity and immediately report cancellation to the workflow

data ActivityTimeoutPolicy

Constructors

StartToCloseTimeout !Duration

ScheduleToCloseTimeout !Duration

StartToCloseAndScheduleToCloseTimeout !Duration !Duration

class StartActivityTimeoutOption a where

Methods

toStartActivityTimeoutOption :: a -> ActivityTimeoutPolicy

defaultStartActivityOptions :: StartActivityTimeoutOption timeout => timeout -> StartActivityOptions

startActivity :: (RequireCallStack, ActivityRef activity) => activity -> StartActivityOptions -> ActivityArgs activity :->: Workflow (Task (ActivityResult activity))

executeActivity :: (RequireCallStack, ActivityRef activity) => activity -> StartActivityOptions -> ActivityArgs activity :->: Workflow (ActivityResult activity)

data StartLocalActivityOptions

Constructors

StartLocalActivityOptions

Fields

activityId :: Maybe ActivityId

scheduleToCloseTimeout :: Maybe Duration

Indicates how long the caller is willing to wait for local activity completion. Limits how long retries will be attempted. When not specified defaults to the workflow execution timeout (which may be unset).

scheduleToStartTimeout :: Maybe Duration

Limits time the local activity can idle internally before being executed. That can happen if the worker is currently at max concurrent local activity executions. This timeout is always non retryable as all a retry would achieve is to put it back into the same queue. Defaults to schedule_to_close_timeout if not specified and that is set. Must be <= schedule_to_close_timeout when set, otherwise, it will be clamped down.

startToCloseTimeout :: Maybe Duration

Maximum time the local activity is allowed to execute after the task is dispatched. This timeout is always retryable. Either or both of schedule_to_close_timeout and this must be specified. If set, this must be <= schedule_to_close_timeout, otherwise, it will be clamped down.

retryPolicy :: Maybe RetryPolicy

Specify a retry policy for the local activity. By default local activities will be retried indefinitely.

localRetryThreshold :: Maybe Duration

If the activity is retrying and backoff would exceed this value, lang will be told to schedule a timer and retry the activity after. Otherwise, backoff will happen internally in core. Defaults to 1 minute.

cancellationType :: ActivityCancellationType

Defines how the workflow will wait (or not) for cancellation of the activity to be confirmed. Lang should default this to WAIT_CANCELLATION_COMPLETED, even though proto will default to TRY_CANCEL automatically.

headers :: Map Text Payload

defaultStartLocalActivityOptions :: StartLocalActivityOptions

startLocalActivity :: (RequireCallStack, ActivityRef act) => act -> StartLocalActivityOptions -> ActivityArgs act :->: Workflow (Task (ActivityResult act))

Child workflow operations

A Child Workflow Execution is a Workflow Execution that is spawned from within another Workflow.

A Workflow Execution can be both a Parent and a Child Workflow Execution because any Workflow can spawn another Workflow.

data StartChildWorkflowOptions

Constructors

StartChildWorkflowOptions

Fields

cancellationType :: ChildWorkflowCancellationType

parentClosePolicy :: ParentClosePolicy

timeoutOptions :: TimeoutOptions

retryPolicy :: Maybe RetryPolicy

cronSchedule :: Maybe Text

initialMemo :: Map Text Payload

searchAttributes :: Map SearchAttributeKey SearchAttributeType

headers :: Map Text Payload

workflowIdReusePolicy :: WorkflowIdReusePolicy

workflowId :: Maybe WorkflowId

taskQueue :: Maybe TaskQueue

defaultChildWorkflowOptions :: StartChildWorkflowOptions

class WorkflowRef f where

Associated Types

type WorkflowArgs f :: [Type]

type WorkflowResult f

Methods

workflowRef :: f -> KnownWorkflow (WorkflowArgs f) (WorkflowResult f)

startChildWorkflow :: (RequireCallStack, WorkflowRef wf) => wf -> StartChildWorkflowOptions -> WorkflowArgs wf :->: Workflow (ChildWorkflowHandle (WorkflowResult wf))

Start a child Workflow execution

Returns a client-side handle that implements a child Workflow interface.

By default, a child will be scheduled on the same task queue as its parent.

A child Workflow handle supports awaiting completion, signaling and cancellation via the returned handle.

In order to query the child, use a WorkflowClient from an Activity.

executeChildWorkflow :: (RequireCallStack, WorkflowRef wf) => wf -> StartChildWorkflowOptions -> WorkflowArgs wf :->: Workflow (WorkflowResult wf)

data ChildWorkflowHandle result

A client side handle to a single child Workflow instance.

It can be used to signal, wait for completion, and cancel the workflow.

class Wait h where

Some tasks in a Workflow return a handle that can be used to wait for the task to complete.

This class provides a common interface for waiting on these handles.

Associated Types

type WaitResult h

Methods

wait :: h -> WaitResult h

Wait for a handle on an an action to complete.

class Cancel h where

Some tasks in a Workflow return a handle that can be used to cancel a task.

This class provides a common interface for performing cancellation on these handles.

Associated Types

type CancelResult h

Methods

cancel :: h -> CancelResult h

Signal to Temporal that a handle representing an async action should be cancelled.

class WorkflowHandle (h :: k -> Type) where

A client side handle to a single Workflow instance. It can be used to signal a workflow execution.

Given the following Workflow definition:

Methods

signal :: forall ref (result :: k). (SignalRef ref, RequireCallStack) => h result -> ref -> SignalArgs ref :->: Workflow (Task ())

Signal a running Workflow.

waitChildWorkflowResult :: RequireCallStack => ChildWorkflowHandle result -> Workflow result

waitChildWorkflowStart :: RequireCallStack => ChildWorkflowHandle result -> Workflow ()

cancelChildWorkflowExecution :: RequireCallStack => ChildWorkflowHandle result -> Workflow ()

data ExternalWorkflowHandle result

Handle representing an external Workflow Execution.

This handle can only be cancelled and signalled.

To call other methods, like query and result, use a WorkflowClient.getHandle inside an Activity.

getExternalWorkflowHandle :: RequireCallStack => WorkflowId -> Maybe RunId -> Workflow (ExternalWorkflowHandle result)

Returns a client-side handle that can be used to signal and cancel an existing Workflow execution. It takes a Workflow ID and optional run ID.

data Info

Constructors

Info

Fields

historyLength :: !Word32

attempt :: !Int

continuedRunId :: !(Maybe RunId)

cronSchedule :: !(Maybe Text)

executionTimeout :: !(Maybe Duration)

headers :: !(Map Text Payload)

namespace :: !Namespace

parent :: !(Maybe ParentInfo)

rawMemo :: !(Map Text Payload)

retryPolicy :: !(Maybe RetryPolicy)

runId :: !RunId

runTimeout :: !(Maybe Duration)

searchAttributes :: !(Map SearchAttributeKey SearchAttributeType)

startTime :: !SystemTime

taskQueue :: !TaskQueue

taskTimeout :: !Duration

workflowId :: !WorkflowId

workflowType :: !WorkflowType

continueAsNewSuggested :: !Bool

data RetryPolicy

A Retry Policy is a collection of attributes that instructs the Temporal Server how to retry a failure of a Workflow Execution or an Activity Task Execution.

Constructors

RetryPolicy

Fields

initialInterval :: Duration

Amount of time that must elapse before the first retry occurs.

The default value is 1 second.

Use case: This is used as the base interval time for the Backoff Coefficient to multiply against.

backoffCoefficient :: Double

The value dictates how much the retry interval increases.

The default value is 2.0.

A backoff coefficient of 1.0 means that the retry interval always equals the Initial Interval.

Use case: Use this attribute to increase the interval between retries. By having a backoff coefficient greater than 1.0, the first few retries happen relatively quickly to overcome intermittent failures, but subsequent retries happen farther and farther apart to account for longer outages. Use the Maximum Interval attribute to prevent the coefficient from increasing the retry interval too much.

maximumInterval :: Maybe Duration

Specifies the maximum interval between retries.

The default value is 100x of the Initial Interval.

Use case: This attribute is useful for Backoff Coefficients that are greater than 1.0 because it prevents the retry interval from growing infinitely.

maximumAttempts :: Int32

Specifies the maximum number of execution attempts that can be made in the presence of failures.

The default is unlimited.

If this limit is exceeded, the execution fails without retrying again. When this happens an error is returned.

Setting the value to 0 also means unlimited.

Setting the value to 1 means a single execution attempt and no retries.

Setting the value to a negative integer results in an error when the execution is invoked.

Use case: Use this attribute to ensure that retries do not continue indefinitely. However, in the majority of cases, we recommend relying on the Workflow Execution Timeout, in the case of Workflows, or Schedule-To-Close Timeout, in the case of Activities, to limit the total duration of retries instead of using this attribute.

nonRetryableErrorTypes :: Vector Text

Description: Specifies errors that shouldn't be retried.

The default is an empty vector. Errors are matched against the type field of the Application Failure. If one of those errors occurs, a retry does not occur.

defaultRetryPolicy :: RetryPolicy

data ParentInfo

Constructors

ParentInfo

Fields

parentNamespace :: Namespace

parentRunId :: RunId

parentWorkflowId :: WorkflowId

Workflow metadata

info :: Workflow Info

We recommend calling info whenever accessing Info fields. Some Info fields change during the lifetime of an Execution— like historyLength and searchAttributes— and some may be changeable in the future— like taskQueue.

getMemoValues :: Workflow (Map Text Payload)

Current workflow's raw memo values.

lookupMemoValue :: Text -> Workflow (Maybe Payload)

Lookup a memo value by key.

upsertSearchAttributes :: RequireCallStack => Map SearchAttributeKey SearchAttributeType -> Workflow ()

Updates this Workflow's Search Attributes by merging the provided searchAttributes with the existing Search Attributes

Using this function will overwrite any existing Search Attributes with the same key.

Versioning workflows

Versioning (known as "patching" in the Haskell library) lets you update Workflow Definitions without causing non-deterministic behavior in current long-running Workflows.

You may need to patch if:

You want to change the remaining logic of a Workflow while it is still running

If your new logic can result in a different execution path

patched :: RequireCallStack => PatchId -> Workflow Bool

Patch or upgrade workflow code by checking or stating that this workflow has a certain patch.

See official Temporal docs page for info.

If the workflow is replaying an existing history, then this function returns true if that history was produced by a worker which also had a patched call with the same patchId.

If the history was produced by a worker without such a call, then it will return false.

If the workflow is not currently replaying, then this call always returns true.

Your workflow code should run the "new" code if this returns true, if it returns false, you should run the "old" code. By doing this, you can maintain determinism.

deprecatePatch :: RequireCallStack => PatchId -> Workflow ()

Indicate that a patch is being phased out.

See official Temporal docs page for info.

Workflows with this call may be deployed alongside workflows with a patched call, but they must not be deployed while any workers still exist running old code without a patched call, or any runs with histories produced by such workers exist. If either kind of worker encounters a history produced by the other, their behavior is undefined.

Once all live workflow runs have been produced by workers with this call, you can deploy workers which are free of either kind of patch call for this ID. Workers with and without this call may coexist, as long as they are both running the "new" code.

Concurrency within Workflows

race :: RequireCallStack => Workflow a -> Workflow b -> Workflow (Either a b)

This function takes two Workflow computations as input, and returns the output of whichever computation finished first. The evaluation of the remaining computation is discontinued.

Note: This function doesn't have a way to explicitly signal to the incomplete other computation that it will no longer be evaluated, so you should be careful about ensuring that incomplete computations are not problematic for your problem domain.

race_ :: RequireCallStack => Workflow a -> Workflow b -> Workflow ()

Run two Workflow actions concurrently, and return the first to finish.

Unlike 'Control.Concurrent.Async.race, this function doesn't explicitly cancel the other computation. If you want to cancel the other computation, you should return sufficient context to do so manually

concurrently :: Workflow a -> Workflow b -> Workflow (a, b)

Run two Workflow actions concurrently, and return both results. If either action throws an exception at any time, the other action will run to completion as long as the exception is caught and handled.

concurrently_ :: Workflow a -> Workflow b -> Workflow ()

Like concurrently, but ignore the result values.

mapConcurrently :: Traversable t => (a -> Workflow b) -> t a -> Workflow (t b)

Map an effect-performing function over over any Traversable data type, performing each action concurrently, returning the original data structure with the arguments replaced with the results.

This is actually a bit of a misnomer, since it's really traverseConcurrently, but this is copied to mimic the async package's naming to slightly ease adoption.

mapConcurrently_ :: Foldable t => (a -> Workflow b) -> t a -> Workflow ()

Alias for traverseConcurrently_

replicateConcurrently :: Int -> Workflow a -> Workflow [a]

Perform the action in the given number of evaluation branches.

replicateConcurrently_ :: Int -> Workflow a -> Workflow ()

Perform the action in the given number of evaluation branches, discarding the results

forConcurrently :: Traversable t => t a -> (a -> Workflow b) -> Workflow (t b)

traverseConcurrently with the arguments flipped.

forConcurrently_ :: Foldable t => t a -> (a -> Workflow b) -> Workflow ()

traverseConcurrently_ with the arguments flipped.

traverseConcurrently :: Traversable t => (a -> Workflow b) -> t a -> Workflow (t b)

Evaluate the action in the given number of evaluation branches, accumulating the results

traverseConcurrently_ :: Foldable t => (a -> Workflow b) -> t a -> Workflow ()

sequenceConcurrently :: Traversable t => t (Workflow a) -> Workflow (t a)

Evaluate each Workflow action in the structure concurrently, and collect the results.

sequenceConcurrently_ :: Foldable t => t (Workflow a) -> Workflow ()

Evaluate each Workflow action in the structure concurrently, and ignore the results.

newtype ConcurrentWorkflow a

A value of type ConcurrentWorkflow a is a Workflow operation that can be composed with other ConcurrentWorkflow values, using the Applicative and Alternative instances.

ConcurrentWorkflow effectively operates as a computation DAG. It's evaluated by exploring the expression across all of the branches of computation until every branch becomes blocked waiting on results from Temporal.

Once that happens, it submits all commands that it has accumulated to the Temporal orchestrator and waits to be reactivated with commands from the orchestrator that unblock computation on one or more branches. This is repeated until the computation completes. When Applicative is used, that means that each computation branch doesn't depend on the result of the others, so they can all explore as far as possible and get blocked. If we use Monad, then that means that the result of the previous computation has to become unblocked in order to provide the result to the subsequent computation.

Constructors

ConcurrentWorkflow

Fields

runConcurrentWorkflowActions :: Workflow a

independently :: Workflow a -> Workflow b -> Workflow (a, b)

biselect :: RequireCallStack => Workflow (Either a b) -> Workflow (Either a c) -> Workflow (Either a (b, c))

biselectOpt :: RequireCallStack => (l -> Either a b) -> (r -> Either a c) -> (a -> t) -> ((b, c) -> t) -> Workflow l -> Workflow r -> Workflow t

The biselectOpt function combines two workflows and applies optimized selection logic.

This function is inspired by the Haxl library's Haxl.Core.Parallel functions. It takes two workflows and combines them, applying discrimination functions to their results to determine the final outcome. The function works as follows:

It explores both workflows concurrently.

If the left workflow completes first: - If the first argument returns Left, the result is immediately returned using the third function. - If the first argument returns Right, it waits for the right workflow to complete.

If the right workflow completes first: - If the second argument returns Left, the result is immediately returned using the third function. - If the second argument returns Right, it waits for the left workflow to complete.

If both workflows complete: - The results are combined using the fourth function if both discriminators return Right.

If either workflow throws an exception, the exception is propagated.

If either workflow is blocked, the function manages the blocking and resumption of computation.

This function optimizes the execution by short-circuiting when possible and managing concurrency efficiently. Be cautious when using this function, as exceptions and evaluation order can be unpredictable depending on the discriminator functions provided.

Interacting with running Workflows

Queries

A Query is a synchronous operation that is used to get the state of a Workflow Execution. The state of a running Workflow Execution is constantly changing. You can use Queries to expose the internal Workflow Execution state to the external world. Queries are available for running or completed Workflows Executions only if the Worker is up and listening on the Task Queue.

Queries are strongly consistent and are guaranteed to return the most recent state. This means that the data reflects the state of all confirmed Events that came in before the Query was sent. An Event is considered confirmed if the call creating the Event returned success. Events that are created while the Query is outstanding may or may not be reflected in the Workflow state the Query result is based on.

A Query can carry arguments to specify the data it is requesting. And each Workflow can expose data to multiple types of Queries.

A Query cannot mutate the state of the Workflow Execution— that is, Queries are read-only and cannot contain any blocking code. This means, for example, that Query handling logic cannot schedule Activity Executions.

class QueryRef query where

Associated Types

type QueryArgs query :: [Type]

type QueryResult query

Methods

queryRef :: query -> KnownQuery (QueryArgs query) (QueryResult query)

data Query a

The Query monad is a very constrained version of the Workflow monad. It can only read state variables and return values. It is used to define query handlers.

data KnownQuery (args :: [Type]) result

Queries are sent from a Temporal Client to a Workflow Execution. The API call is synchronous. The Query is identified at both ends by a Query name. The Workflow must have a Query handler that is developed to handle that Query and provide data that represents the state of the Workflow Execution.

Queries are strongly consistent and are guaranteed to return the most recent state. This means that the data reflects the state of all confirmed Events that came in before the Query was sent. An Event is considered confirmed if the call creating the Event returned success. Events that are created while the Query is outstanding may or may not be reflected in the Workflow state the Query result is based on.

A Query can carry arguments to specify the data it is requesting. And each Workflow can expose data to multiple types of Queries.

A Query must never mutate the state of the Workflow Execution— that is, Queries are read-only and cannot contain any blocking code. This means, for example, that Query handling logic cannot schedule Activity Executions.

Sending Queries to completed Workflow Executions is supported, though Query reject conditions can be configured per Query

Constructors

(Codec codec result, ApplyPayloads codec args, GatherArgs codec args) => KnownQuery

Fields

queryName :: Text

queryCodec :: codec

setQueryHandler :: (QueryRef query, f ~ (QueryArgs query :->: Query (QueryResult query)), RequireCallStack) => query -> f -> Workflow ()

Register a query handler.

The handler will be called when a query with the given name is received.

Signals

class SignalRef sig where

Associated Types

type SignalArgs sig :: [Type]

Methods

signalRef :: sig -> KnownSignal (SignalArgs sig)

data KnownSignal (args :: [Type])

Constructors

(ApplyPayloads codec args, GatherArgs codec args) => KnownSignal

Fields

signalName :: Text

signalCodec :: codec

setSignalHandler :: (ValidSignalHandler f, RequireCallStack, SignalRef ref, ArgsOf f ~ SignalArgs ref) => ref -> f -> Workflow ()

type ValidSignalHandler f = (ResultOf Workflow f ~ (), (ArgsOf f :->: Workflow ()) ~ f)

data Condition a

A very restricted Monad that allows for reading StateVar values. This Monad keeps track of which StateVar values are read so that it can block and retry the computation if any of the values change.

waitCondition :: RequireCallStack => Condition Bool -> Workflow ()

Wait on a condition to become true before continuing.

This must be used with signals, steps executed concurrently via the Applicative instance, or with the race command, as those are the only way for state to change in a workflow while a portion of the workflow itself is in this blocking condition.

N.B. this should be used with care, as it can lead to the workflow suspending indefinitely if the condition is never met. (e.g. if there is no signal handler that changes the state appropriately)

Other utilities

unsafeAsyncEffectSink :: RequireCallStack => IO () -> Workflow ()

While workflows are deterministic, there are categories of operational concerns (metrics, logging, tracing, etc.) that require access to IO operations like the network or filesystem. The IO monad is not generally available in the Workflow monad, but you can use sink to run an IO action in a workflow. In order to maintain determinism, the operation will be executed asynchronously and does not return a value. Be sure that the sink operation terminates, or else you will leak memory and/or threads.

Do not use sink for any Workflow logic, or else you will violate determinism.

State vars

data StateVar a

StateVar values are mutable variables scoped to a Workflow run.

Workflows are deterministic, so you may not use normal IORefs, since the IORef could have been created outside of the workflow and cause nondeterminism.

However, it is totally safe to mutate state variables as long as they are scoped to a workflow and derive their state transitions from the workflow's deterministic execution.

StateVar values may also be read from within a query and mutated within signal handlers.

newStateVar :: a -> Workflow (StateVar a)

class MonadReadStateVar (m :: Type -> Type) where

Methods

readStateVar :: StateVar a -> m a

class MonadWriteStateVar (m :: Type -> Type) where

Methods

writeStateVar :: StateVar a -> a -> m ()

modifyStateVar :: StateVar a -> (a -> a) -> m ()

Time and timers

now :: Workflow UTCTime

Current time from the workflow perspective.

The time returned is updated only when the workflow performs an operation in the Workflow monad that blocks. Examples of such operations are sleep, awaitCondition, awaitActivity, awaitWorkflow, etc.

Equivalent to getCurrentTime from the time package.

time :: RequireCallStack => Workflow SystemTime

Current time from the workflow perspective.

The value is relative to epoch time.

sleep :: RequireCallStack => Duration -> Workflow ()

data Timer

createTimer :: Duration -> Workflow (Maybe Timer)

Asynchronous sleep.

Creates a timer that fires after the specified duration.

The timer is not guaranteed to fire immediately after the duration expires, but it is intended to fire as close to the expiration as possible.

Note that the timer is started when the command is received by the Temporal Platform, not when the timer is created. The command is sent as soon as the workflow is blocked by any operation, such as sleep, awaitCondition, awaitActivity, awaitWorkflow, etc.

If the duration is less than or equal to zero, the timer will not be created.

scheduledTime :: Workflow (Maybe UTCTime)

When using the Schedules feature of Temporal, this is the effective time that an instance was scheduled to run, according to the TemporalScheduledStartTime search attribute.

This is useful for backfilling or retrying a scheduled Workflow if you need the schedule to have a stable start time.

Workflow cancellation

isCancelRequested :: RequireCallStack => Workflow Bool

Workflows may be sent a cancellation request from the Temporal Platform, but Workflow code is not required to respond to the cancellation request.

In order to opt in to cancellation handling, you can call isCancelRequested periodically within your workflow code to check whether a cancellation request has been received.

waitCancellation :: RequireCallStack => Workflow ()

Block the current workflow's main execution thread until the workflow is cancelled.

The workflow can still respond to signals and queries while waiting for cancellation.

Upon cancellation, the workflow will throw a WorkflowCancelRequested exception.

This function is useful for actor-style workflows that perform work in response to signals and/or respond to queries, but otherwise need to remain idle on their main codepath.

N.B. It is likely not safe to call this in a signal handler.

Random value generation

Workflow executions are deterministic, so you can't use the usual IO-based random number generation.

Instead, each workflow execution is given a seed from the Temporal platform to seed a PRNG. This allows you to generate random values in a deterministic way.

The Workflow monad provides a RandomGen instance, so you can use the usual random and randomR functions from the Random and Stateful modules.

randomGen :: Workflow WorkflowGenM

Get a mutable randomness generator for the workflow.

uuid4 :: Workflow UUID

Generate an RFC compliant V4 uuid.

Uses the workflow's deterministic PRNG, making it safe for use within a workflow.

This function is cryptographically insecure.

uuid7 :: RequireCallStack => Workflow UUID

Generates a UUIDv7 using the current time (from time) and random data (from workflowRandomnessSeed).

data WorkflowGenM

Continue as new

data ContinueAsNewOptions

Constructors

ContinueAsNewOptions

Fields

taskQueue :: Maybe TaskQueue

runTimeout :: Maybe Duration

taskTimeout :: Maybe Duration

retryPolicy :: Maybe RetryPolicy

memo :: Map Text Payload

searchAttributes :: Map SearchAttributeKey SearchAttributeType

headers :: Map Text Payload

defaultContinueAsNewOptions :: ContinueAsNewOptions

continueAsNew

Arguments

:: WorkflowRef wf

=> wf
The workflow to continue as new. It doesn't have to be the same as the current workflow.

-> ContinueAsNewOptions

-> WorkflowArgs wf :->: Workflow (WorkflowResult wf)

Continue-As-New is a mechanism by which the latest relevant state is passed to a new Workflow Execution, with a fresh Event History.

As a precautionary measure, the Temporal Platform limits the total Event History to 51,200 Events or 50 MB, and will warn you after 10,240 Events or 10 MB. To prevent a Workflow Execution Event History from exceeding this limit and failing, use Continue-As-New to start a new Workflow Execution with a fresh Event History.

All values passed to a Workflow Execution through parameters or returned through a result value are recorded into the Event History. A Temporal Cluster stores the full Event History of a Workflow Execution for the duration of a Namespace's retention period. A Workflow Execution that periodically executes many Activities has the potential of hitting the size limit.

A very large Event History can adversely affect the performance of a Workflow Execution. For example, in the case of a Workflow Worker failure, the full Event History must be pulled from the Temporal Cluster and given to another Worker via a Workflow Task. If the Event history is very large, it may take some time to load it.

The Continue-As-New feature enables developers to complete the current Workflow Execution and start a new one atomically.

The new Workflow Execution has the same Workflow Id, but a different Run Id, and has its own Event History. TODO, don't make this an exception, make it a return value

Type definitions

type GatherArgs codec (args :: [Type]) = (VarArgs args, AllArgs (Codec codec) args)

newtype ActivityId

A unique identifier for an Activity Execution.

Constructors

ActivityId

Fields

rawActivityId :: Text

newtype WorkflowId

A Workflow Id is a customizable, application-level identifier for a Workflow Execution that is unique to an Open Workflow Execution within a Namespace.

Constructors

WorkflowId

Fields

rawWorkflowId :: Text

newtype Namespace

A Namespace is a unit of isolation within the Temporal Platform

Constructors

Namespace

Fields

rawNamespace :: Text

newtype TaskQueue

A Task Queue is a queue that a Worker Process polls for Tasks.

Task Queues do not have any ordering guarantees. It is possible to have a Task that stays in a Task Queue for a period of time, if there is a backlog that wasn't drained for that time.

Task Queues are very lightweight components. Task Queues do not require explicit registration but instead are created on demand when a Workflow Execution or Activity spawns or when a Worker Process subscribes to it. When a Task Queue is created, both a Workflow Task Queue and an Activity Task Queue are created under the same name. There is no limit to the number of Task Queues a Temporal Application can use or a Temporal Cluster can maintain.

Constructors

TaskQueue

Fields

rawTaskQueue :: Text

newtype PatchId

Constructors

PatchId

Fields

rawPatchId :: Text

newtype RunId

A Run Id is a globally unique, platform-level identifier for a Workflow Execution.

The current Run Id is mutable and can change during a Workflow Retry. You shouldn't rely on storing the current Run Id, or using it for any logical choices, because a Workflow Retry changes the Run Id and can lead to non-determinism issues.

Temporal guarantees that only one Workflow Execution with a given Workflow Id can be in an Open state at any given time. But when a Workflow Execution reaches a Closed state, it is possible to have another Workflow Execution in an Open state with the same Workflow Id. For example, a Temporal Cron Job is a chain of Workflow Executions that all have the same Workflow Id. Each Workflow Execution within the chain is considered a Run.

A Run Id uniquely identifies a Workflow Execution even if it shares a Workflow Id with other Workflow Executions.

Constructors

RunId

Fields

rawRunId :: Text

data ParentClosePolicy

Used by the service to determine the fate of a child workflow in case its parent is closed.

Constructors

ParentClosePolicyUnspecified
Lets the server set the default.

ParentClosePolicyTerminate
Terminate the child workflow.

ParentClosePolicyAbandon
Do not terminate the child workflow. The child workflow continues to run.

ParentClosePolicyRequestCancel
Request cancellation on the child workflow.

data ChildWorkflowCancellationType

Controls at which point to report back when a child workflow is cancelled.

Constructors

ChildWorkflowCancellationAbandon
Do not request cancellation of the child workflow if already scheduled

ChildWorkflowCancellationTryCancel
Initiate a cancellation request and immediately report cancellation to the parent workflow

ChildWorkflowCancellationWaitCancellationCompleted
Wait for child cancellation completion.

ChildWorkflowCancellationWaitCancellationRequested
Request cancellation of the child and wait for confirmation that the request was received.

data WorkflowIdReusePolicy

A Workflow Id Reuse Policy determines whether a Workflow Execution is allowed to spawn with a particular Workflow Id, if that Workflow Id has been used with a previous, and now Closed, Workflow Execution.

It is not possible for a new Workflow Execution to spawn with the same Workflow Id as another Open Workflow Execution, regardless of the Workflow Id Reuse Policy. In some cases, an attempt to spawn a Workflow Execution with a Workflow Id that is the same as the Id of a currently Open Workflow Execution results in a Workflow execution already started error.

Constructors

WorkflowIdReusePolicyUnspecified

WorkflowIdReusePolicyAllowDuplicate
The Workflow Execution is allowed to exist regardless of the Closed status of a previous Workflow Execution with the same Workflow Id. This is currently the default policy, if one is not specified. Use this when it is OK to have a Workflow Execution with the same Workflow Id as a previous, but now Closed, Workflow Execution.

WorkflowIdReusePolicyAllowDuplicateFailedOnly
The Workflow Execution is allowed to exist only if a previous Workflow Execution with the same Workflow Id does not have a Completed status. Use this policy when there is a need to re-execute a Failed, Timed Out, Terminated or Cancelled Workflow Execution and guarantee that the Completed Workflow Execution will not be re-executed.

WorkflowIdReusePolicyRejectDuplicate
The Workflow Execution cannot exist if a previous Workflow Execution has the same Workflow Id, regardless of the Closed status. Use this when there can only be one Workflow Execution per Workflow Id within a Namespace for the given retention period.

WorkflowIdReusePolicyTerminateIfRunning
Specifies that if a Workflow Execution with the same Workflow Id is already running, it should be terminated and a new Workflow Execution with the same Workflow Id should be started. This policy allows for only one Workflow Execution with a specific Workflow Id to be running at any given time.

newtype WorkflowType

This is generally the name of the function itself

Constructors

WorkflowType

Fields

rawWorkflowType :: Text

type RequireCallStack = (HasCallStack, RequireCallStackImpl)

This constraint is similar to HasCallStack in that it's presence will capture a stack frame for the call site of the function. Unlike HasCallStack, this is not a "magic" constraint that is automagically solved by GHC, which means that calling a function with this constraint will cause GHC to ask you to add this constraint to your own function.

For example, let's say you have a function unsafeHead :: RequireCallStack => [a] -> a. Then you go to call that function:

myCoolFunction :: [Int] -> Int
myCoolFunction = unsafeHead

GHC will complain about the lack of the RequireCallStack constraint. You will have two options:

Add the constraint to your functions. This is a good option because it means the callstack from unsafeHead will include the myCoolFunction callsite as well.

   myCoolFunction :: RequireCallStack => [Int] -> Int
   myCoolFunction = unsafeHead
   

Use provideCallStack to silence the error. This will truncate the callstack unless you use HasCallStack above. You should only do this if you're confident that you don't need any debugging information from a more complete callstack.

   myCoolFunction :: [Int] -> Int
   myCoolFunction = provideCallStack unsafeHead
   

Since: require-callstack-0.1.0.0

data TimeoutOptions

Constructors

TimeoutOptions

Fields

executionTimeout :: Maybe Duration

A Workflow Execution Timeout is the maximum time that a Workflow Execution can be executing (have an Open status) including retries and any usage of Continue As New.

The default value is ∞ (infinite). If this timeout is reached, the Workflow Execution changes to a Timed Out status. This timeout is different from the Workflow Run Timeout. This timeout is most commonly used for stopping the execution of a Temporal Cron Job after a certain amount of time has passed.

runTimeout :: Maybe Duration

A Workflow Run Timeout is the maximum amount of time that a single Workflow Run is restricted to.

The default is set to the same value as the Workflow Execution Timeout. This timeout is most commonly used to limit the execution time of a single Temporal Cron Job Execution.

If the Workflow Run Timeout is reached, the Workflow Execution is Terminated.

taskTimeout :: Maybe Duration

A Workflow Task Timeout is the maximum amount of time allowed for a Worker to execute a Workflow Task after the Worker has pulled that Workflow Task from the Task Queue.

The default value is 10 seconds. This timeout is primarily available to recognize whether a Worker has gone down so that the Workflow Execution can be recovered on a different Worker. The main reason for increasing the default value would be to accommodate a Workflow Execution that has a very long Workflow Execution History that could take longer than 10 seconds for the Worker to load.

defaultTimeoutOptions :: TimeoutOptions

Commonly used

type family (args :: [Type]) :->: result where ... infixr 0

Construct a function type from a list of argument types and a result type.

Equations

('[] :: [Type]) :->: result = result

(arg ': args) :->: result = arg -> args :->: result